Search View Done
Finished the search view page, there is one known bug and it is the category dropdown menu button.
I think due to the way that the bootstrap dropdown menu class is defined and the way I nested rows and divs
to create the desired look when the button is clicked, it may or may not render the dropdown menu correctly. Sometimes it
appears at the bottom of the screen. I used ngClass directive to insert/remove the dropdown-menu class which is causing
the problem. When it is removed only the dropdown class is left which renders it view as if the menu is opened, when the class
dropdown-menu is inserted back it renders the view as if it closed, however this is when the bug may or may not occur. To get rid of the bug
temporarily I've noticed that if one clicks/highlights any text it will remove the bug.