Continued working on Registration View
Added more fields to the view. Attempted to create dropdown menu for the
security questions but for some reason unknown to me the first button is binding
the the second one. I added the same logic from search-view.component.ts to the register.
component.ts for the toggle effect. I added all the fields in one page because eventually
I will add a button at the top with an ngIf to determine what field will pop up
or not.