Finally was able to interpolate nested json
Finally was able to interpolate nested json property from response.
I interpolated the business description, and address fields.

In business-entity-component.ts I removed the type for the response so
I could access the promise body (JSON) data property and assign it to
this.business. In order to overcome angular complaining about the
address field being null I assigned the address property to another
object in the component.ts and interpolated the address valus via that
address object.