Interpolated business username on click event
When the name of the businesses are clicked in the search view, their
business username is passed as an argument to the redirectTo() function
and console.log() is used to print the username to the console.
Changed the html tag of the business name to a span and created custom
css for it.